added active models tab

+++ visualization/experiments/views.py
@@ -11,6 +11,28 @@ from shutil import copyfile
 from django.core.files import File
 from visualization.settings import MEDIA_ROOT
 from datasets.models import ImageModel
+from torch_model_module.models import TorchModel
+
+
+from torch_datasets.flask_api import *
+
+def Sortifier(img_path_list):
+    import requests
+    url = "http://localhost:5000/sortifier"
+
+    if isinstance(img_path_list,str):
+        img_path_list = [img_path_list]
+
+    pred_dict = {}
+    for img_path in img_path_list:
+        #try:
+        resp = requests.post(url, files={"file": open(img_path,'rb')}).json() 
+        print(resp)    
+        pred_dict.update({img_path:resp['class_name']})
+        #except:
+        #print(f'Something went wrong when predicting {img_path}') 
+
+    return pred_dict
 
 def get_images(images_path):
     dest_path = MEDIA_ROOT
@@ -66,12 +88,37 @@ def detail_experiment(request, ex_id):
     if experiment:
         img_list = experiment.dataset.img_list.all()
 
-    context = {'experiment': experiment,'img_list':img_list}
+    active_models = TorchModel.objects.filter(active = True)
+
+    context = {'experiment': experiment,'img_list':img_list,'active_models':active_models}
     if request.method == 'POST':
+        if 'back' in request.POST:
+            return redirect('/experiments/')
         if 'delete' in request.POST:
             ExperimentModel.objects.filter(id=ex_id).delete()
             return redirect('/experiments/')
+        if 'activate' in request.POST:
+            #create api
+            torch_model = experiment.torch_model
+            model_name = torch_model.name
+            if torch_model.active == False:
+                torch_model.active = True
+                torch_model.save()
+                context.update({'message':f'model {model_name} activated'})
+            else:
+                context.update({'message':f'model {model_name} already active'})
 
+        for model in active_models:
+            if f'deactivate_{model.id}' in request.POST:
+                model.active = False
+                model.save()
+                context.update({'message':f'model {model.name} deactivated'})
+
+        
+        active_models = TorchModel.objects.filter(active = True)
+        context.update({'active_models':active_models})
+
+ 
     return render(request, 'experiments/detail_experiment.html', context)
 
 
@@ -81,16 +128,34 @@ def detail_image(request, ex_id,img_id):
     
     context = {'img': img,'experiment':experiment}
     if request.method == 'POST':
+        if 'predict' in request.POST:
+            class_index_dict = {0:'bitewings',1:'cephalometrics',2:'panoramics',3:'periapicals'}
+            model_path = 'torch_datasets/model_checkpoint.pth'
+            model_name = 'resnet34'
+
+            #model = load_model(model_name,model_path,class_index_dict)
+            #app = define_app(model)
+            import subprocess
+            from subprocess import call, run
+            run(["python", "torch_datasets/flask_api.py"], stdout=subprocess.DEVNULL)
+            print('flask app running')
+            prediction_dict = Sortifier('torch_datasets/tmp.jpg')
+            
+            #app.run(host='localhost', port=5050,debug=True,use_reloader=False)
+            
+            message = f'{prediction_dict}'
+            context.update({'message':message})
+            return render(request, 'experiments/detail_image.html', context)
         if 'XAI' in request.POST:
             message = 'XAI should happen'
             context.update({'message':message})
             return render(request, 'experiments/detail_image.html', context)
-        elif 'refresh' in request.POST:
+        if 'refresh' in request.POST:
             context = {'img': img}
             return render(request, 'experiments/detail_image.html', context)
-        elif 'board' in request.POST:
+        if 'board' in request.POST:
             return redirect(f'/experiments/{ex_id}')
-        elif 'delete' in request.POST:
+        if 'delete' in request.POST:
             ImageModel.objects.filter(id=img_id).delete()
             return redirect(f'/experiments/{ex_id}')
